+++ src/Twig/Extension/AppExtension.php
@@ -13,6 +13,7 @@ class AppExtension extends AbstractExtension
         return [
             new TwigFunction('get_media_embed_code', [AppExtensionRuntime::class, 'getMediaEmbedCode']),
             new TwigFunction('get_media_templates', [AppExtensionRuntime::class, 'getMediaTemplates']),
+            new TwigFunction('format_value_with_unit', [AppExtensionRuntime::class, 'formatValueWithUnit']),
         ];
     }
 }