+++ CHANGELOG.md
@@ -11,6 +11,9 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ## [Unreleased]
 
+- [PR-39](https://github.com/itk-dev/aapodwalk_api/pull/39)
+  Changed distance and duration to integer values.
+  Added "Value with unit" field
 - [PR-38](https://github.com/itk-dev/aapodwalk_api/pull/38)
   Added mailer config
 - [PR-34](https://github.com/itk-dev/aapodwalk_api/pull/34)